+++ src/bulkscan_helpers.jl
@@ -1,6 +1,7 @@
 struct Results_by_bin
     idxs_by_bin::Array{Array{Bool, 1}, 1}
     LODs_by_bin::Array{Array{Float64, 2}, 1}
+    Effect_sizes_by_bin::Array{Array{Float64, 2}, 1}
     h2_taken::Array{Float64, 1};
 end
 
@@ -46,18 +47,22 @@ Inputs are rotated, re-weighted.
 """
 function computeR_LMM(wY::Array{Float64, 2}, wX::Array{Float64, 2}, wIntercept::Array{Float64, 2})
 
-    # exclude the effect of (rotated) intercept (idea is similar as centering data in the linear model case)
+    # Exclude the effect of (rotated) intercept (idea is similar as centering data in the linear model case)
     Y00 = resid(wY, wIntercept);
     X00 = resid(wX, wIntercept);
 
-    # standardize the response and covariates by dividing by their norms
+    # Standardize the response and covariates by dividing by their norms
     norm_Y = mapslices(x -> norm(x), Y00, dims = 1) |> vec;
     norm_X = mapslices(x -> norm(x), X00, dims = 1) |> vec;
+
+    # replace!(norm_Y, 0 => 1.0)
+    # replace!(norm_X, 0 => 1.0)
 
     colDivide!(Y00, norm_Y);
     colDivide!(X00, norm_X);
 
-    R = X00' * Y00; # p-by-m matrix
+    # Matrix of correlation coefficients between the trait and all markers
+    R = X00' * Y00;
 
     return R
 
@@ -127,25 +132,34 @@ Assumes the heritabilities only differ by traits but remain the same across all 
 function univar_liteqtl(y0_j::AbstractArray{Float64, 1}, X0_intercept::AbstractArray{Float64, 2}, 
                         X0_covar::AbstractArray{Float64, 2}, lambda0::AbstractArray{Float64, 1}; 
                         prior_variance = 0.0, prior_sample_size = 0.0,
-                        reml::Bool = false, optim_interval::Int64 = 1)
+                        reml::Bool = false, optim_interval::Int64 = 1
+                        )
 
     n = size(y0_j, 1);
     y0 = reshape(y0_j, :, 1);
 
-    # estimate the heritability from the null model and apply it to the reweighting of all markers;
+    # Estimate the heritability from the null model
     vc = fitlmm(y0, X0_intercept, lambda0, [prior_variance, prior_sample_size]; 
                 reml = reml, optim_interval = optim_interval);
+    # Construct weights to adjust for heteroskedasticity due to heritability
     sqrtw = sqrt.(abs.(makeweights(vc.h2, lambda0)));
 
-    # re-weight the data; then in theory, the observations are homoskedestic and independent.
+    # Re-weight the data: final transformed data wy0 are homoskedestic and independent
     wy0 = rowMultiply(y0, sqrtw);
     wX0_intercept = rowMultiply(X0_intercept, sqrtw);
     wX0_covar = rowMultiply(X0_covar, sqrtw);
 
+    # Matrix of correlation coefficients between the trait and all markers
     R = computeR_LMM(wy0, wX0_covar, wX0_intercept);
+
+    # Estimate effect sizes for baseline covariates and markers
+    wX0_covar_intercept = hcat(wX0_intercept, wX0_covar);
+    B = wX0_covar_intercept\wy0; # effect sizes for all markers + intercept
+
     threaded_map!(r2lod, R, n; dims = 2);
 
-    return (R = R, h2 = vc.h2); # results will be p-by-1, i.e. all LOD scores for the j-th trait and p markers
+    # Return LOD scores, effect sizes, and heritability estimate
+    return (B = B, R = R, h2 = vc.h2);
 
 end
 
@@ -183,6 +197,7 @@ function weighted_liteqtl(Y0::Array{Float64, 2}, X0::Array{Float64, 2},
 
     wY0 = rowMultiply(Y0, sqrtw);
     wX0 = rowMultiply(X0, sqrtw);
+    b = wX0 \ wY0;
 
     if num_of_covar == 1
         wX0_intercept = reshape(wX0[:, 1], :, 1);
@@ -197,7 +212,7 @@ function weighted_liteqtl(Y0::Array{Float64, 2}, X0::Array{Float64, 2},
 
     threaded_map!(r2lod, LOD, n; dims = dims); # results will be p-by-1, i.e. all LOD scores for the j-th trait and p markers
 
-    return LOD
+    return (LOD = LOD, B = b)
 end
 
 ## GridBulk helper functions:
@@ -249,7 +264,8 @@ function gridscan_by_bin(pheno::Array{Float64, 2}, geno::Array{Float64, 2},
 
     # Y_std = colStandardize(pheno);
     Y_std = pheno;
-
+    ##################################################################################
+    ## Step 1: Rotate to decorrelate the individuals: This is the same for all traits
     (Y0, X0, lambda0) = transform_rotation(Y_std, [covar geno], kinship; 
                                            addIntercept = addIntercept, decomp_scheme = decomp_scheme);
 
@@ -264,6 +280,9 @@ function gridscan_by_bin(pheno::Array{Float64, 2}, geno::Array{Float64, 2},
 
     X0_intercept = X0[:, 1:num_of_covar];
 
+    ##################################################################################
+    ## Step 2: Perform a quick scan of h2's (under null and by grid-search) for all traits
+    
     weights_each_h2 = map(x -> makeweights(x, lambda0), grid); # make weights evaluated on each h2 in grid
     ell_results = map(x -> wls_multivar(Y0, X0_intercept, x, prior; reml = reml).Ell, weights_each_h2);
     ell_results = reduce(vcat, ell_results);
@@ -278,45 +297,71 @@ function gridscan_by_bin(pheno::Array{Float64, 2}, geno::Array{Float64, 2},
     h2_taken = unique(values(idxs_sets));
     nbins = length(h2_taken);
 
+    ##################################################################################
+    ## Step 3: Distribute traits by h2 values into bins: Traits inside a bin share the same h2
+    
     blocking_idxs = distribute_traits_by_h2(idxs_sets, h2_taken, m, nbins);
+    # LOD scores for each bin
     results = Array{Array{Float64, 2}, 1}(undef, nbins);
+    # Effect sizes for each bin
+    effect_sizes_by_bin = Array{Array{Float64, 2}, 1}(undef, nbins);
+    
 
-    ## Threads.@threads for t in 1:nbins
+    ##################################################################################
+    ## Step 4: Compute LOD scores for each bin of traits using the weighted LiteQTL approach
+    
     for t in 1:nbins
-        results[t] = weighted_liteqtl(Y0[:, blocking_idxs[t]], X0, lambda0, h2_taken[t]; 
+        out = weighted_liteqtl(Y0[:, blocking_idxs[t]], X0, lambda0, h2_taken[t]; 
                                       num_of_covar = num_of_covar);
+
+        results[t] = out.LOD;
+        # selected_effects = vcat(1, (num_of_covar+1):size(out.B, 1)) # baseline (1) + marker (num_of_covar+1:end) effects
+        selected_effects = (num_of_covar+1):size(out.B, 1) # marker effects only
+        effect_sizes_by_bin[t] = out.B[selected_effects, :]; # exclude the covariate effects
+        # print(size(effect_sizes_by_bin[t]))
+
     end
 
-    return Results_by_bin(blocking_idxs, results, h2_taken)
+    ##################################################################################
+    ## Final output: LOD scores for traits ordered by the order of bins
+    ## (Downstream to reorder the results by the original trait order)
+    return Results_by_bin(blocking_idxs, results, effect_sizes_by_bin, h2_taken)
     
 end
 
 function reorder_results(blocking_idxs::Array{Array{Bool, 1}, 1}, 
-                         lods_by_block::Array{Array{Float64, 2}, 1}, 
+                         lods_by_block::Array{Array{Float64, 2}, 1},
+                         effect_sizes_by_block::Array{Array{Float64, 2}, 1}, 
                          m::Int64, p::Int64)
     
     LOD = Array{Float64, 2}(undef, p, m);
+    B = Array{Float64, 2}(undef, p, m);
     
     
     for block in 1:length(blocking_idxs)
         idxs_curr_block = blocking_idxs[block];
         LOD[:, idxs_curr_block] = lods_by_block[block];
+        B[:, idxs_curr_block] = effect_sizes_by_block[block];
     end
     
-    return LOD
+    return (LOD = LOD, B = B)
     
 end
 
 ## MaxBulk helper functions:
 """
-tmax!(max, toCompare)
+tmax!(currL1, nextL1, currB, nextB, h2_panel, h2_panel_counter, h2_list)
 
 Does element-wise comparisons of two 2d Arrays and keep the larger elements in-place. 
 
 # Arguments
-- max = 2d Array of Float; matrix of current maximum values
-- toCompare = 2d Array of Flopat; matrix of values to compare with the current maximum values
-- 
+- currL1 = 2d Array of Float; matrix of current maximum l1 (alt-loglik) values over h2 grid
+- nextL1 = 2d Array of Float; matrix of values to compare with the current maximum values
+- currB = 2d Array of Float; matrix to store the effect sizes corresponding to the current maximum l1 values
+- nextB = 2d Array of Float; matrix of effect sizes corresponding to the next L1 values
+- h2_panel = 2d Array of Float; matrix to store the optimal h2 values for each entry
+- h2_panel_counter = 2d Array of Int; matrix to store the indices of the optimal h2 values for each entry
+- h2_list = 1d Array of Float; list of candidate h2 values
 
 # Value
 
@@ -324,23 +369,32 @@ Nothing; does in-place maximizations.
 
 # Notes:
 
-Will modify input matrix `max` by a parallelized loop; uses @tturbo in the package `LoopVectorization.jl`
+Updates optimal values in `currL1` in-place, in a threaded loop.
 
 """
-function tmax!(max::Array{Float64, 2}, toCompare::Array{Float64, 2},
-               hsq_panel::Array{Float64, 2}, hsq_panel_counter::Array{Int64, 2},
-               hsq_list::Array{Float64, 1})
+function tmax!(currL1::Array{Float64, 2}, nextL1::Array{Float64, 2},
+               currB::Array{Float64, 2}, nextB::Array{Float64, 2},  
+               h2_panel::Array{Float64, 2}, 
+               h2_panel_counter::Array{Int64, 2},
+               h2_list::Array{Float64, 1})
     
-    (p, m) = size(max);
+    (p, m) = size(currL1);
     
-    Threads.@threads for j in 1:m
-        for i in 1:p
+    Threads.@threads for j in 1:m # Multiprocessing over traits over threads
+        for i in 1:p # Loop over markers
             
-            # change inputs in-place only when condition is met (the coordinate of LOD is higher on the current h2 value)
-            if (max[i, j] < toCompare[i, j])
-                max[i, j] = toCompare[i, j];
-                hsq_panel_counter[i, j] = hsq_panel_counter[i, j]+1;
-                hsq_panel[i, j] = hsq_list[hsq_panel_counter[i, j]];
+            # Update each l1(i, j) value in-place, if current l1(i, j) is smaller:
+            if (currL1[i, j] < nextL1[i, j])
+
+                # Update new maximum LOD score:
+                currL1[i, j] = nextL1[i, j];
+
+                # Record the new optimal h2 value:
+                h2_panel_counter[i, j] = h2_panel_counter[i, j]+1; # Index of the h2 value in h2_list
+                h2_panel[i, j] = h2_list[h2_panel_counter[i, j]]; # Update the optimal h2 value from indices
+
+                # Also, update the effect sizes based on the new optimal l1:
+                currB[i, j] = nextB[i, j];
             end
 
             # do nothing if not.